fix: simplify button click check in shouldEnhanceNavigation
Remove unnecessary `.toShort()` conversion in button comparison.
The event.button property is already a number type in JS, so the
direct comparison with 0 is cleaner and more idiomatic.

+++ mosaik-docs-client/src/jsMain/kotlin/mosaik/docs/client/Main.kt
@@ -85,7 +85,7 @@ fun shouldEnhanceNavigation(
     }
 
     // Don't enhance if it's not a left-click (button 0 is left-click)
-    if (event.button != 0.toShort()) {
+    if (event.button != 0) {
         return false
     }
